Extract PartitionedOutputInfo as a standalone class
