+++ presto-main/src/main/java/com/facebook/presto/operator/PartitionedOutputOperator.java
@@ -26,9 +26,6 @@ import com.facebook.presto.spi.block.RunLengthEncodedBlock;
 import com.facebook.presto.spi.plan.PlanNodeId;
 import com.facebook.presto.spi.relation.ConstantExpression;
 import com.facebook.presto.spi.type.Type;
-import com.facebook.presto.util.Mergeable;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.util.concurrent.ListenableFuture;
 import io.airlift.units.DataSize;
 
@@ -40,7 +37,6 @@ import java.util.function.Function;
 
 import static com.facebook.presto.execution.buffer.PageSplitterUtil.splitPage;
 import static com.facebook.presto.spi.block.PageBuilderStatus.DEFAULT_MAX_PAGE_SIZE_IN_BYTES;
-import static com.google.common.base.MoreObjects.toStringHelper;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static java.lang.Math.max;
 import static java.lang.Math.min;
@@ -442,68 +438,6 @@ public class PartitionedOutputOperator
                     rowsAdded.addAndGet(pagePartition.getPositionCount());
                 }
             }
-        }
-    }
-
-    public static class PartitionedOutputInfo
-            implements Mergeable<PartitionedOutputInfo>, OperatorInfo
-    {
-        private final long rowsAdded;
-        private final long pagesAdded;
-        private final long outputBufferPeakMemoryUsage;
-
-        @JsonCreator
-        public PartitionedOutputInfo(
-                @JsonProperty("rowsAdded") long rowsAdded,
-                @JsonProperty("pagesAdded") long pagesAdded,
-                @JsonProperty("outputBufferPeakMemoryUsage") long outputBufferPeakMemoryUsage)
-        {
-            this.rowsAdded = rowsAdded;
-            this.pagesAdded = pagesAdded;
-            this.outputBufferPeakMemoryUsage = outputBufferPeakMemoryUsage;
-        }
-
-        @JsonProperty
-        public long getRowsAdded()
-        {
-            return rowsAdded;
-        }
-
-        @JsonProperty
-        public long getPagesAdded()
-        {
-            return pagesAdded;
-        }
-
-        @JsonProperty
-        public long getOutputBufferPeakMemoryUsage()
-        {
-            return outputBufferPeakMemoryUsage;
-        }
-
-        @Override
-        public PartitionedOutputInfo mergeWith(PartitionedOutputInfo other)
-        {
-            return new PartitionedOutputInfo(
-                    rowsAdded + other.rowsAdded,
-                    pagesAdded + other.pagesAdded,
-                    Math.max(outputBufferPeakMemoryUsage, other.outputBufferPeakMemoryUsage));
-        }
-
-        @Override
-        public boolean isFinal()
-        {
-            return true;
-        }
-
-        @Override
-        public String toString()
-        {
-            return toStringHelper(this)
-                    .add("rowsAdded", rowsAdded)
-                    .add("pagesAdded", pagesAdded)
-                    .add("outputBufferPeakMemoryUsage", outputBufferPeakMemoryUsage)
-                    .toString();
         }
     }
 }
